Resolve conflicts in logger.c
Adapt to the changes in the logger and store (manager)

+++ clients/odbc/driver/ODBCConvert.c
@@ -3,7 +3,7 @@
  * License, v. 2.0.  If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  *
- * Copyright 2008-2015 MonetDB B.V.
+ * Copyright 1997 - July 2008 CWI, August 2008 - 2016 MonetDB B.V.
  */
 
 #include "ODBCGlobal.h"
@@ -511,7 +511,7 @@ parseoptionalbracketednumber(char **svalp,
 			return SQL_ERROR;
 		slen -= (int) (eptr - sval);
 		sval = eptr;
-		*val2p = val;
+		*val2p = (int) val;
 	}
 
 	if (slen == 0 || *sval != ')')
